# platform-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hyperfleet Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /accounts:
    post:
      summary: Enable an account
      description: |
        Enables an AWS account for Cedar/AVP-based authorization.
        Creates a policy store in Amazon Verified Permissions.
        Requires privileged access.
      operationId: createAccount
      tags:
        - Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableAccountRequest"
      responses:
        "201":
          description: Account enabled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Account already enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List accounts
      description: |
        Returns a list of all enabled accounts.
        Requires privileged access.
      operationId: listAccounts
      tags:
        - Authorization
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountList"
        "403":
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /accounts/{id}:
    get:
      summary: Get an account
      description: |
        Returns details of a specific enabled account.
        Requires privileged access.
      operationId: getAccount
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: AWS account ID
          schema:
            type: string
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Disable an account
      description: |
        Disables an AWS account from authorization.
        Deletes the policy store and all associated policies, groups, and attachments.
        Requires privileged access.
      operationId: deleteAccount
      tags:
        - Authorization
      parameters:
        - name: id
          in: path
          required: true
          description: AWS account ID
          schema:
            type: string
      responses:
        "204":
          description: Account disabled successfully
        "403":
          description: Forbidden - caller is not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Authorization - Check
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - kind
        - code
        - reason
      properties:
        kind:
          type: string
          description: Resource kind
          example: Error
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
...
    EnableAccountRequest:
      type: object
      description: Request body for enabling an account
      required:
        - accountId
      properties:
        accountId:
          type: string
          description: AWS account ID to enable
        privileged:
          type: boolean
          description: If true, account bypasses all authorization checks
          default: false
    Account:
      type: object
      description: An enabled account
      required:
        - accountId
        - privileged
        - createdAt
      properties:
        accountId:
          type: string
          description: AWS account ID
          pattern: '^\d{12}$'
        policyStoreId:
          type: string
          description: AVP policy store ID (null if privileged)
        privileged:
          type: boolean
          description: If true, bypasses all authorization
        createdAt:
          type: string
          format: date-time
          description: When the account was enabled
        createdBy:
          type: string
          description: ARN of who enabled the account
    AccountList:
      type: object
      description: List of accounts
      required:
        - kind
        - items
        - total
      properties:
        kind:
          type: string
          example: AccountList
        items:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        total:
          type: integer
          description: Total number of accounts
...